+++ codegen/src/main.rs
@@ -94,6 +94,8 @@ fn generate_content(items: Vec<Item>) -> TokenStream {
         //!
         //! Modify the source that source for control on what to include and how.
 
+        #![allow(deprecated)]
+
         #![cfg_attr(feature = #PY_FEATURE, allow(unsafe_op_in_unsafe_fn))]
 
         #[cfg(feature = #PY_FEATURE)]